+++ src/tripCard/index.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { color, radius, space, componentSizes } from '_utils/branding'
+import { color, radius, space, componentSizes, transition } from '_utils/branding'
 
 import TripCard from './TripCard'
 
@@ -8,9 +8,10 @@ const StyledTripCard = styled(TripCard)`
     border-radius: ${radius.l};
     box-shadow: 0 1pt 4pt rgba(0, 0, 0, 0.16), 0 2pt 8pt rgba(0, 0, 0, 0.08);
     list-style-type: none;
+    transition: box-shadow ${transition.duration.base} ${transition.easing.default};
   }
 
-  .kirk-tripCard:hover {
+  &:hover {
     box-shadow: 0 2pt 8pt rgba(0, 0, 0, 0.08), 0 2pt 16pt rgba(0, 0, 0, 0.08);
   }